requirements.txt
deepagents>=0.3.0
langgraph>=1.0.5
langchain-openai>=1.1.5
tavily-python>=0.7.17
python-dotenv>=1.2.1
langsmith>=0.1.0